Make filedialog test run
Also need to update FileDialog some and the test to
avoid failures.
